+++ android/src/main/kotlin/com/rohit/ble_peripheral/BlePeripheralPlugin.kt
@@ -50,6 +50,7 @@ class BlePeripheralPlugin : FlutterPlugin, BlePeripheralChannel, ActivityAware {
     private val emptyBytes = byteArrayOf()
     private val listOfDevicesWaitingForBond = mutableListOf<String>()
     private var isAdvertising: Boolean? = null
+    private var requireBonding: Boolean = true
 
     override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
         BlePeripheralChannel.setUp(flutterPluginBinding.binaryMessenger, this)
@@ -83,11 +84,7 @@ class BlePeripheralPlugin : FlutterPlugin, BlePeripheralChannel, ActivityAware {
 
     override fun isSupported(): Boolean {
         val bluetoothAdapter = bluetoothManager?.adapter ?: return false
-        // if (!bluetoothAdapter.isEnabled) throw UnsupportedOperationException("Bluetooth is disabled.")
-        if (!bluetoothAdapter.isMultipleAdvertisementSupported) throw UnsupportedOperationException(
-            "Bluetooth LE Advertising not supported on this device."
-        )
-        return true
+        return bluetoothAdapter.isMultipleAdvertisementSupported
     }
 
     override fun addService(service: BleService) {
@@ -110,20 +107,28 @@ class BlePeripheralPlugin : FlutterPlugin, BlePeripheralChannel, ActivityAware {
         } ?: emptyList()
     }
 
+    override fun getSubscribedClients(): List<SubscribedClient> {
+        return subscribedCharDevicesMap.map { data ->
+            SubscribedClient(data.key, data.value)
+        }
+    }
+
     override fun startAdvertising(
         services: List<String>,
         localName: String?,
         timeout: Long?,
         manufacturerData: ManufacturerData?,
         addManufacturerDataInScanResponse: Boolean,
+        requireBonding: Boolean,
     ) {
+        this.requireBonding = requireBonding
         if (!isBluetoothEnabled()) {
             enableBluetooth()
             throw Exception("Bluetooth is not enabled")
         }
 
         handler?.post { // set up advertising setting
-            localName?.let { bluetoothManager?.adapter?.name = it }
+            bluetoothManager?.adapter?.name = localName ?: Build.MODEL
             val advertiseSettings = AdvertiseSettings.Builder()
                 .setTxPowerLevel(AdvertiseSettings.ADVERTISE_TX_POWER_HIGH)
                 .setConnectable(true)
@@ -235,6 +240,12 @@ class BlePeripheralPlugin : FlutterPlugin, BlePeripheralChannel, ActivityAware {
         }
     }
 
+    private fun onBonded(device: BluetoothDevice) {
+        synchronized(bluetoothDevicesMap) {
+            bluetoothDevicesMap[device.address] = device
+        }
+    }
+
     private fun cleanConnection(device: BluetoothDevice) {
         val deviceAddress = device.address
 
@@ -290,27 +301,22 @@ class BlePeripheralPlugin : FlutterPlugin, BlePeripheralChannel, ActivityAware {
                 super.onConnectionStateChange(device, status, newState)
                 when (newState) {
                     BluetoothProfile.STATE_CONNECTED -> {
-                        if (device.bondState == BluetoothDevice.BOND_NONE) {
+                        if (requireBonding && device.bondState == BluetoothDevice.BOND_NONE) {
+                            Log.d(TAG, "Device not bonded. Trying to bond")
                             // Wait for bonding
                             listOfDevicesWaitingForBond.add(device.address)
                             device.createBond()
-                        } else if (device.bondState == BluetoothDevice.BOND_BONDED) {
-                            handler?.post {
-                                gattServer?.connect(device, true)
-                            }
-                            synchronized(bluetoothDevicesMap) {
-                                bluetoothDevicesMap.put(
-                                    device.address,
-                                    device
-                                )
-                            }
+                        } else if (!requireBonding || device.bondState == BluetoothDevice.BOND_BONDED) {
+                            onBonded(device)
                         }
                         onConnectionUpdate(device, status, newState)
                     }
 
                     BluetoothProfile.STATE_DISCONNECTED -> {
                         val deviceAddress = device.address
-                        synchronized(bluetoothDevicesMap) { bluetoothDevicesMap.remove(deviceAddress) }
+                        synchronized(bluetoothDevicesMap) {
+                            bluetoothDevicesMap.remove(deviceAddress)
+                        }
                         onConnectionUpdate(device, status, newState)
                     }
 
@@ -486,11 +492,17 @@ class BlePeripheralPlugin : FlutterPlugin, BlePeripheralChannel, ActivityAware {
                         val charList: MutableList<String> =
                             subscribedCharDevicesMap[it] ?: mutableListOf()
                         if (isSubscribed) {
-                            charList.add(characteristicId)
+                            if (!charList.contains(characteristicId)) {
+                                charList.add(characteristicId)
+                            }
                         } else if (charList.contains(characteristicId)) {
                             charList.remove(characteristicId)
                         }
-                        subscribedCharDevicesMap[it] = charList
+                        if (charList.isEmpty()) {
+                            subscribedCharDevicesMap.remove(it)
+                        } else {
+                            subscribedCharDevicesMap[it] = charList
+                        }
                     }
 
                 }
@@ -569,7 +581,7 @@ class BlePeripheralPlugin : FlutterPlugin, BlePeripheralChannel, ActivityAware {
 
                 val state =
                     intent.getIntExtra(BluetoothDevice.EXTRA_BOND_STATE, BluetoothDevice.ERROR)
-                val device: BluetoothDevice? =
+                val device: BluetoothDevice =
                     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
                         intent.getParcelableExtra(
                             BluetoothDevice.EXTRA_DEVICE,
@@ -578,23 +590,31 @@ class BlePeripheralPlugin : FlutterPlugin, BlePeripheralChannel, ActivityAware {
                     } else {
                         @Suppress("DEPRECATION")
                         intent.getParcelableExtra(BluetoothDevice.EXTRA_DEVICE)
-                    }
+                    } ?: return
 
                 handler?.post {
                     bleCallback?.onBondStateChange(
-                        device?.address ?: "",
+                        device.address,
                         state.toBondState(),
                     ) {}
                 }
 
-                // if waiting for connection and device is bonded
-                val waitingForConnection = listOfDevicesWaitingForBond.contains(device?.address)
-                if (state == BluetoothDevice.BOND_BONDED && device != null && waitingForConnection) {
-                    listOfDevicesWaitingForBond.remove(device.address)
-                    handler?.post {
-                        gattServer?.connect(device, true)
+                if (listOfDevicesWaitingForBond.contains(device.address)) {
+                    if (bluetoothManager?.getConnectionState(
+                            device,
+                            BluetoothProfile.GATT
+                        ) == BluetoothGatt.STATE_CONNECTED
+                    ) {
+                        onBonded(device)
+                    } else {
+                        handler?.post {
+                            Log.d(TAG, "Trying to connect again")
+                            gattServer?.connect(device, false)
+                        }
                     }
                 }
+
+                listOfDevicesWaitingForBond.removeAll { it == device.address }
             }
         }
     }