+++ lib/src/pigeon/ble_peripheral_pigeon.dart
@@ -76,6 +76,12 @@ class BlePeripheralPigeon extends BlePeripheralInterface {
   Future<List<String>> getServices() async =>
       (await _channel.getServices()).whereType<String>().toList();
 
+    @override
+    Future<List<SubscribedClient>> getSubscribedClients() async =>
+      (await _channel.getSubscribedClients())
+        .whereType<SubscribedClient>()
+        .toList();
+
   /// To update the value of a characteristic
   @override
   Future<void> updateCharacteristic({
@@ -95,6 +101,7 @@ class BlePeripheralPigeon extends BlePeripheralInterface {
     int? timeout,
     ManufacturerData? manufacturerData,
     bool addManufacturerDataInScanResponse = false,
+    bool requireBonding = true,
   }) {
     if (defaultTargetPlatform == TargetPlatform.windows &&
         manufacturerData == null) {
@@ -105,7 +112,7 @@ class BlePeripheralPigeon extends BlePeripheralInterface {
       );
     }
     return _channel.startAdvertising(services, localName, timeout,
-        manufacturerData, addManufacturerDataInScanResponse);
+      manufacturerData, addManufacturerDataInScanResponse, requireBonding);
   }
 
   /// Stop advertising